check for online and return

+++ src/api/provider.js
@@ -1,3 +1,6 @@
+import nanoid from 'nanoid';
+import Task from '../models/task.js';
+
 export default class Provider {
   constructor(api, store) {
     this._api = api;
@@ -5,19 +8,38 @@ export default class Provider {
   }
 
   getTasks() {
-    return this._api.getTasks();
+    if (this._isOnline()) {
+      return this._api.getTasks();
+    }
+
+    return Promise.resolve(Task.parseTasks([]));
   }
 
   createTask(task) {
-    return this._api.createTask(task);
+    if (this._isOnline()) {
+      return this._api.createTask(task);
+    }
+
+    const fakeNewTaskId = nanoid();
+    const fakeNewTask = Task.parseTask(Object.assign({}, task.toRAW(), {id: fakeNewTaskId}));
+
+    return Promise.resolve(fakeNewTask);
   }
 
   updateTask(id, task) {
-    return this._api.updateTask(id, task);
+    if (this._isOnline()) {
+      return this._api.updateTask(id, task);
+    }
+
+    return Promise.resolve(task);
   }
 
   deleteTask(id) {
-    return this._api.deleteTask(id);
+    if (this._isOnline()) {
+      return this._api.deleteTask(id);
+    }
+
+    return Promise.resolve();
   }
 
   _isOnline() {
